+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.2.0
 ------
 
 Added `Debug` attribute
+Added TOML and YAML support
 Temporarily switched to `github.com/runeimp/gotenv` from `github.com/subosito/gotenv`